Moved to an API generated with SpeakEasy (#21)

+++ lib/hashcat/types.go
@@ -17,42 +17,38 @@ type StatusGuess struct {
 	GuessBase        string  `json:"guess_base"`         // The base wordlist used for the attack
 	GuessBaseCount   int64   `json:"guess_base_count"`   // The number of words in the base wordlist
 	GuessBaseOffset  int64   `json:"guess_base_offset"`  // The offset into the base wordlist
-	GuessBasePercent float32 `json:"guess_base_percent"` // The percentage of the base wordlist that has been processed
-
-	GuessMod        string  `json:"guess_mod"`         // The modifier wordlist used for the attack
-	GuessModCount   int64   `json:"guess_mod_count"`   // The number of words in the modifier wordlist
-	GuessModOffset  int64   `json:"guess_mod_offset"`  // The offset into the modifier wordlist
-	GuessModPercent float32 `json:"guess_mod_percent"` // The percentage of the modifier wordlist that has been processed
-
-	GuessMode int32 `json:"guess_mode"` // The attack mode
+	GuessBasePercent float64 `json:"guess_base_percent"` // The percentage of the base wordlist that has been processed
+	GuessMod         string  `json:"guess_mod"`          // The modifier wordlist used for the attack
+	GuessModCount    int64   `json:"guess_mod_count"`    // The number of words in the modifier wordlist
+	GuessModOffset   int64   `json:"guess_mod_offset"`   // The offset into the modifier wordlist
+	GuessModPercent  float64 `json:"guess_mod_percent"`  // The percentage of the modifier wordlist that has been processed
+	GuessMode        int64   `json:"guess_mode"`         // The attack mode
 }
 
 type StatusDevice struct {
-	DeviceID   int32  `json:"device_id"`   // The device ID
+	DeviceID   int64  `json:"device_id"`   // The device ID
 	DeviceName string `json:"device_name"` // The device name
 	DeviceType string `json:"device_type"` // The device type
 	Speed      int64  `json:"speed"`       // The speed of the device
-	Util       int32  `json:"util"`        // The utilization of the device
-	Temp       int32  `json:"temp"`        // The temperature of the device
+	Util       int64  `json:"util"`        // The utilization of the device
+	Temp       int64  `json:"temp"`        // The temperature of the device
 }
 
 type Status struct {
-	OriginalLine string    `json:"original_line"` // The original line from hashcat
-	Time         time.Time `json:"time"`          // The time the status was received
-
+	OriginalLine    string         `json:"original_line"`    // The original line from hashcat
+	Time            time.Time      `json:"time"`             // The time the status was received
 	Session         string         `json:"session"`          // The session ID
 	Guess           StatusGuess    `json:"guess"`            // The current guess
-	Status          int32          `json:"status"`           // The status of the attack
+	Status          int64          `json:"status"`           // The status of the attack
 	Target          string         `json:"target"`           // The target hash
 	Progress        []int64        `json:"progress"`         // The progress of the attack
 	RestorePoint    int64          `json:"restore_point"`    // The restore point
-	RecoveredHashes []int32        `json:"recovered_hashes"` // The number of recovered hashes
-	RecoveredSalts  []int32        `json:"recovered_salts"`  // The number of recovered salts
+	RecoveredHashes []int64        `json:"recovered_hashes"` // The number of recovered hashes
+	RecoveredSalts  []int64        `json:"recovered_salts"`  // The number of recovered salts
 	Rejected        int64          `json:"rejected"`         // The number of rejected hashes
 	Devices         []StatusDevice `json:"devices"`          // The devices used for the attack
-
-	TimeStart     int64 `json:"time_start"`     // The start time of the attack
-	EstimatedStop int64 `json:"estimated_stop"` // The estimated stop time of the attack
+	TimeStart       int64          `json:"time_start"`       // The start time of the attack
+	EstimatedStop   int64          `json:"estimated_stop"`   // The estimated stop time of the attack
 }
 
 type Result struct {
